refactor(settings): give the signature-verification screen a route
_SignatureVerificationPolicyScreen was private to
nostr_settings_screen.dart and reached by Navigator.push, so it had no
path. It takes no constructor arguments, so nothing blocked routing it.

Extract it as SignatureVerificationPolicyScreen in its own file and
register it as a child of /nostr-settings, mirroring Nip05SettingsScreen
— that parent already has a routes: block, so this joins the existing
nested pattern rather than introducing one.

Being a real child route, cold entry keeps /nostr-settings in the match
list, so back has somewhere to go. It still uses safePop for the same
reason every other routed back button does.

The two policy-copy helpers move with it and become public: the hub tile
also renders the subtitle, so duplicating them would have been the only
alternative.
