+++ mobile/lib/router/routes/settings_routes.dart
@@ -33,6 +33,7 @@ import 'package:openvine/screens/settings/monetization_links_settings_screen.dar
 import 'package:openvine/screens/settings/nip05_settings_screen.dart';
 import 'package:openvine/screens/settings/nostr_settings_screen.dart';
 import 'package:openvine/screens/settings/settings_screen.dart';
+import 'package:openvine/screens/settings/signature_verification_policy_screen.dart';
 import 'package:openvine/screens/settings/storage/storage_management_page.dart';
 import 'package:openvine/screens/settings/support_center_screen.dart';
 import 'package:openvine/screens/settings/supporter_screen.dart';
@@ -167,6 +168,11 @@ List<RouteBase> settingsRoutes(Ref ref) {
           path: Nip05SettingsScreen.subpath,
           name: Nip05SettingsScreen.routeName,
           builder: (_, _) => const Nip05SettingsScreen(),
+        ),
+        GoRoute(
+          path: SignatureVerificationPolicyScreen.subpath,
+          name: SignatureVerificationPolicyScreen.routeName,
+          builder: (_, _) => const SignatureVerificationPolicyScreen(),
         ),
       ],
     ),
